- many GUI improvements - refactoring - new raw stack saving with one file per stack

+++ src/clearcontrol/devices/cameras/devices/sim/test/StackCameraDeviceSimulatorTests.java
@@ -67,7 +67,7 @@ public class StackCameraDeviceSimulatorTests
     });
 
     lStackCameraDeviceSimulator.getExposureInSecondsVariable()
-                               .set((double) 0.001);
+                               .set(0.001);
 
     lStackCameraDeviceSimulator.open();
